+++ src/commands/init_config.rs
@@ -152,6 +152,7 @@ mod tests {
                         command: S("tools/test_2.sh"),
                     },
                 ]),
+                commands: None,
                 applications: Some(ApplicationSection {
                     actionlint: Some(ApplicationNoFile {
                         enabled: Some(true),